Improved seat massager sample
