+++ digital-twin-model/src/lib.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 pub mod sdv_v1;
+pub mod sdv_v2;
 
 use serde_derive::{Deserialize, Serialize};
