+++ samples/command/consumer/src/main.rs
@@ -22,7 +22,7 @@ use uuid::Uuid;
 #[derive(Debug, Serialize, Deserialize)]
 struct ShowNotificationRequestPayload {
     #[serde(rename = "Notification")]
-    notification: String,
+    notification: sdv::hmi::show_notification::request::TYPE,
     #[serde(rename = "$metadata")]
     metadata: Metadata,
 }
