+++ samples/mixed/consumer/src/main.rs
@@ -32,7 +32,7 @@ struct IsAirConditioningActiveProperty {
 #[derive(Debug, Serialize, Deserialize)]
 struct ShowNotificationRequestPayload {
     #[serde(rename = "Notification")]
-    notification: String,
+    notification: sdv::hmi::show_notification::request::TYPE,
     #[serde(rename = "$metadata")]
     metadata: Metadata,
 }
