+++ samples/seat_massager/provider/src/request_impl.rs
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 // SPDX-License-Identifier: MIT
 
-use digital_twin_model::sdv_v1 as sdv;
+use digital_twin_model::sdv_v2 as sdv;
 use log::{debug, info, warn};
 use parking_lot::Mutex;
 use samples_protobuf_data_access::async_rpc::v1::common::Status;
