+++ samples/tutorial/consumer/src/main.rs
@@ -21,7 +21,7 @@ use tonic::Status;
 #[derive(Debug, Serialize, Deserialize)]
 struct ShowNotificationRequestPayload {
     #[serde(rename = "Notification")]
-    notification: String,
+    notification: sdv::hmi::show_notification::request::TYPE,
     #[serde(rename = "$metadata")]
     metadata: Metadata,
 }
